feat(filter): refactor code and add KDoc
